rm dm_plot.c - chance for static functions
